backend/technospace/requirements.txt
dj-database-url==0.5.0
psycopg2-binary==2.9.1
python-decouple==3.4
django-cors-headers==3.12.0
Django==4.1.4
djangorestframework==3.14.0